переделал createFilm

+++ src/main/java/ru/yandex/practicum/filmorate/service/GenreService.java
@@ -19,7 +19,7 @@ public class GenreService {
         return genreDbStorage.findAll();
     }
 
-    public Genre findById(long id) {
+    public Genre findById(int id) {
         try {
             return genreDbStorage.findById(id);
         } catch (DataAccessException e) {
